+++ sdk/eventgrid/azure-messaging-eventgrid/src/main/java/com/azure/messaging/eventgrid/systemevents/AcsMessageEventData.java
@@ -18,8 +18,12 @@ import java.time.format.DateTimeFormatter;
 
 /**
  * Schema of common properties of all chat thread events.
+ * 
+ * @deprecated This class is deprecated and may be removed in future releases. System events are now available in the
+ * azure-messaging-eventgrid-systemevents package.
  */
 @Fluent
+@Deprecated
 public class AcsMessageEventData implements JsonSerializable<AcsMessageEventData> {
 
     /*